Copied CL# 9693447
#jira UE-81409
Bug fixed: Renaming ULayers did not update their internal UActors nor the ViewHiddenLayers that contained them if it was called with a reference to the Layer->LayerName, e.g. RenameLayer(Layer->LayerName, NewLayerName). This created lots of side effects with UActors and HiddenLayers, that after the layer rename pointed to no-longer-existing Layer names.
#rb codereview chris.gagnon

[CL 9693857 by Gines Hidalgo in 4.24 branch]