Get GeometryCollections simulating again.  Lots of caveats as this is an initial checkin of ongoing work...  Assume nothing works other than a single rigid body in a geometry collection falling under gravity.
Adds a particle class for geometry collections derived off of PBDRigidParticle.

Recessitates the geometry collection physics proxy, to the point where it works for the most simple case.  Lots of churn in this class as it had most of it's functionality implemented with an abandoned callback based interface to the solver, and an abandoned rigid body accounting methodology.  This is a work in progress.

#fyi: Brice.Criswell, Michael.Lentine
#rb: Brice.Criswell
#jira: none

[CL 10089537 by Ryan Kautzman in 4.24 branch]