Add (fudged) correct scale for MS roughness lobe.
It was arbitrary taking the TT roughness for evaluating the MS contribution, but usually MS is while more diffuse/high-variance than the actual TT term, even though most of the energy is coming from the TT term. This scale factor (eye-balled) try to compensate for this. This integration needs to be replaced in mid-term.

#rb none
#jira none


#ushell-cherrypick of 10236860 by Charles.deRousiers

[CL 10251727 by Charles deRousiers in 4.24 branch]