The previous frame buffers used as inputs for the sun average and AA postprocess passes on mobile were not transitioned to readable, triggering an ensure on Vulkan.
The buffers were left as readable at the end of the frame, but were transitioned to writable automatically in FRenderTargetPool::TransitionTargetsWritable() at the beginning of the next frame. The sun average input always comes from the sun merge pass, so that can be fixed by setting the target's AutoWritable property to false. The AA input can come from a number of passes depending on how the pipeline is set up, so it's not as easy to figure out which one needs AutoWritable turned off. I've added an explicit transition for now, since this is going to be rewritten for RDG at some point anyway.

#lockdown cristina.riveron
#rb Dmitriy.Dyomin
#jira none

[CL 10808196 by mihnea balta in 4.24 branch]